requirements.txt
asgiref==3.2.10
awscli==1.18.157
boto3==1.15.16
botocore==1.18.16
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1.2
django-ajax-selects==2.0.0
django-appconf==1.0.4
django-heroku==0.3.1
django-imagekit==4.0.2
django-storages==1.10.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
